+++ src/app/interfaces/coffee-user.ts
@@ -2,6 +2,8 @@ export interface User {
     uid: string;
     email: string;
     isAdmin: boolean;
+    emailVerified: boolean;
+    providerId: any;
     firstName?: string;
     lastName?: string;
     displayName?: string;