+++ src/app/pages/home/home.page.ts
@@ -21,28 +21,16 @@ export class HomePage implements OnInit {
   user$: Observable<User>;
 
   bannerConfig: AdMobFreeBannerConfig = {
-    isTesting: false,
-    autoShow: true,
+    isTesting: true,
+    autoShow: false,
     id: environment.admobAppID,
     bannerAtTop: false
   };
 
 
-  constructor(private admob: AdMobFree,
-              private auth: AuthService, 
-              private platform: Platform) {}
+  constructor(private admob: AdMobFree, private auth: AuthService, private platform: Platform) {}
 
-  public async ngOnInit(): Promise<void> {
-    try {
-      this.user$ = this.auth.user$;
-
-      this.admob.banner.config(this.bannerConfig);
-
-      await this.admob.banner.prepare();
-      console.log('banner ready');
-    }
-    catch(e) { console.log('OnInit error: ', e) }
-  }
+  public async ngOnInit(): Promise<void> { this.user$ = this.auth.user$ }
 
   public async googleLogin() {
     try {
@@ -52,5 +40,25 @@ export class HomePage implements OnInit {
       }
     }
     catch(e) { console.log('googleLogin() error: ', e) }
+  }
+
+  private async startAdmob() {
+    try {
+      await this.platform.ready();
+
+      this.admob.banner.config(this.bannerConfig);
+      await this.admob.banner.prepare();
+      await this.admob.banner.show();
+      console.log('banner ready');
+    }
+    catch(e) { console.log('startAdmob() error: ', e) }
+  }
+
+  private async hideAdmob() {
+    try {
+      await this.platform.ready();
+      await this.admob.banner.hide();
+    }
+    catch(e) { console.log('hideAdmob() error: ', e) }
   }
 }